pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ajazz-akp03e"
version = "0.1.0"
description = "Python SDK for the Ajazz AKP03E stream deck"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
authors = [
    { name = "Tomek Ceszke", email = "tomek@ceszke.com" },
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: System :: Hardware",
    "Typing :: Typed",
]
dependencies = [
    "hid>=1.0.6",
    "Pillow>=10.0",
]

[project.optional-dependencies]
dev = [
    "build>=1.0",
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "mypy>=1.10",
    "ruff>=0.5",
]

[project.urls]
Homepage = "https://github.com/tomekceszke/ajazz-akp03e"
Repository = "https://github.com/tomekceszke/ajazz-akp03e"
Issues = "https://github.com/tomekceszke/ajazz-akp03e/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/ajazz_akp03e"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.mypy]
strict = true
packages = ["ajazz_akp03e"]
mypy_path = "src"

[tool.ruff]
target-version = "py311"
src = ["src", "tests"]

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "B", "SIM"]
